Bugfixes
RT Server Messages now have 2 enums added.
MediusTicketLogin upon creating account hardcode a password so DB is happy.
Finalize NetAddress for the BinaryVport
Set Encryption to true by default in case database is not read properly.
Update GetFileMetaData in the DbController

+++ RT.Common/Types.cs
@@ -2859,7 +2859,7 @@ namespace RT.Common
 
     #region DME 
 
-    public enum DME_SERVER_LANGUAGE_TYPE : int
+    public enum DME_SERVER_LANGUAGE_TYPE : byte
     {
          DME_SERVER_LANGUAGE_NONE = 0,
          DME_SERVER_LANGUAGE_US_ENGLISH = 1,
@@ -2876,7 +2876,7 @@ namespace RT.Common
          DME_SERVER_LANGUAGE_TAIWANESE = 12,
     }
 
-    public enum DME_SERVER_ENCODING_TYPE : int
+    public enum DME_SERVER_ENCODING_TYPE : byte
     {
          DME_SERVER_ENCODING_NONE,
          DME_SERVER_ENCODING_ISO_8859_1,
